# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hindsight HTTP API",
    "description": "HTTP API for Hindsight",
    "contact": {
      "name": "Memory System"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/default/banks/{bank_id}/reflect": {
      "post": {
        "summary": "Reflect and generate answer",
        "description": "Reflect and formulate an answer using bank identity, world facts, and opinions.\n\n    This endpoint:\n    1. Retrieves agent facts (bank's identity)\n    2. Retrieves world facts relevant to the query\n    3. Retrieves existing opinions (bank's perspectives)\n    4. Uses LLM to formulate a contextual answer\n    5. Extracts and stores any new opinions formed\n    6. Returns plain text answer, the facts used, and new opinions",
        "operationId": "reflect",
        "parameters": [
          {
            "name": "bank_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bank Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReflectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReflectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Budget": {
        "type": "string",
        "enum": [
          "low",
          "mid",
          "high"
        ],
        "title": "Budget",
        "description": "Budget levels for recall/reflect operations."
      },
...
      "EntityIncludeOptions": {
        "properties": {
          "max_tokens": {
            "type": "integer",
            "title": "Max Tokens",
            "description": "Maximum tokens for entity observations",
            "default": 500
          }
        },
        "type": "object",
        "title": "EntityIncludeOptions",
        "description": "Options for including entity observations in recall results."
      },
...
      "FactsIncludeOptions": {
        "properties": {},
        "type": "object",
        "title": "FactsIncludeOptions",
        "description": "Options for including facts (based_on) in reflect results."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MetadataFilter": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Metadata key to filter on"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value",
            "description": "Value to match. If None with match_unset=True, matches any record where key is not set."
          },
          "match_unset": {
            "type": "boolean",
            "title": "Match Unset",
            "description": "If True, also match records where this metadata key is not set",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "key"
        ],
        "title": "MetadataFilter",
        "description": "Filter for metadata fields. Matches records where (key=value) OR (key not set) when match_unset=True.",
        "example": {
          "key": "source",
          "match_unset": true,
          "value": "slack"
        }
      },
...
      "ReflectFact": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "occurred_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred Start"
          },
          "occurred_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred End"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ReflectFact",
        "description": "A fact used in think response.",
        "example": {
          "context": "healthcare discussion",
          "id": "123e4567-e89b-12d3-a456-426614174000",
          "occurred_end": "2024-01-15T10:30:00Z",
          "occurred_start": "2024-01-15T10:30:00Z",
          "text": "AI is used in healthcare",
          "type": "world"
        }
      },
      "ReflectIncludeOptions": {
        "properties": {
          "facts": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FactsIncludeOptions"
              },
              {
                "type": "null"
              }
            ],
            "description": "Include facts that the answer is based on. Set to {} to enable, null to disable (default: disabled)."
          },
          "entities": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntityIncludeOptions"
              },
              {
                "type": "null"
              }
            ],
            "description": "Include entity observations. Set to {max_tokens: N} to enable, null to disable (default: disabled)."
          }
        },
        "type": "object",
        "title": "ReflectIncludeOptions",
        "description": "Options for including additional data in reflect results."
      },
      "ReflectRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "budget": {
            "$ref": "#/components/schemas/Budget",
            "default": "low"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "filters": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MetadataFilter"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters",
            "description": "Filter by metadata. Multiple filters are ANDed together."
          },
          "include": {
            "$ref": "#/components/schemas/ReflectIncludeOptions",
            "description": "Options for including additional data (both disabled by default)"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "ReflectRequest",
        "description": "Request model for reflect endpoint.",
        "example": {
          "budget": "low",
          "context": "This is for a research paper on AI ethics",
          "filters": [
            {
              "key": "source",
              "match_unset": true,
              "value": "slack"
            }
          ],
          "include": {
            "entities": {
              "max_tokens": 500
            },
            "facts": {}
          },
          "query": "What do you think about artificial intelligence?"
        }
      },
      "ReflectResponse": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "based_on": {
            "items": {
              "$ref": "#/components/schemas/ReflectFact"
            },
            "type": "array",
            "title": "Based On",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ReflectResponse",
        "description": "Response model for think endpoint.",
        "example": {
          "based_on": [
            {
              "id": "123",
              "text": "AI is used in healthcare",
              "type": "world"
            },
            {
              "id": "456",
              "text": "I discussed AI applications last week",
              "type": "agent"
            }
          ],
          "text": "Based on my understanding, AI is a transformative technology..."
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}